# api/openapi/query-api.yaml
openapi: 3.0.3
info:
  title: Query API
  version: 0.2.0
  description: Read-only API backed by PostgreSQL projections maintained by the event processor.
servers:
  - url: http://localhost:8083
paths:
  /health:
    get:
      summary: Readiness and health check
      responses:
        '200':
          description: Service is healthy
  /api/v1/customers:
    get:
      summary: List customers
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: country_code
          schema:
            type: string
        - in: query
          name: segment
          schema:
            type: string
      responses:
        '200':
          description: Paginated customer list
  /api/v1/customers/{customerID}:
    get:
      summary: Get a customer projection
      parameters:
        - in: path
          name: customerID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer found
        '404':
          description: Customer not found
  /api/v1/orders:
    get:
      summary: List orders
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - in: query
          name: customer_id
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Paginated order list
  /api/v1/orders/{orderID}:
    get:
      summary: Get an order projection with items
      parameters:
        - in: path
          name: orderID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order found
        '404':
          description: Order not found
  /api/v1/invoices:
    get:
      summary: List invoices
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - in: query
          name: customer_id
          schema:
            type: string
        - in: query
          name: order_id
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Paginated invoice list
  /api/v1/invoices/{invoiceID}:
    get:
      summary: Get an invoice projection
      parameters:
        - in: path
          name: invoiceID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice found
        '404':
          description: Invoice not found
components:
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0